# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/media/profile-images": {
      "get": {
        "operationId": "Media_listProfileImages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "List the current account profile images",
        "tags": [
          "media"
        ]
      },
      "post": {
        "operationId": "Media_uploadProfileImage",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Upload a profile image for the current account",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/profile-images/{profileImageId}": {
      "delete": {
        "operationId": "Media_deleteProfileImage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Delete one profile image for the current account",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/public/{mediaAssetId}": {
      "get": {
        "operationId": "Media_getPublicMedia",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Serve a public media asset by id",
        "tags": [
          "media"
        ]
      }
    },
...
  },
  "info": {
    "title": "Matcha API",
    "description": "Canonical API contract for the Matcha backend. Target OpenAPI 3.1 where the active toolchain remains smooth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "matcha-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer access token returned by Matcha authentication endpoints."
      }
    },
...
  }
}